pyproject.toml
[project]
name = "buisbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Kotaz",email = "semechkagent@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pyrogram (>=2.0.106,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "google-genai (>=1.11.0,<2.0.0)",
    "tgcrypto-pyrofork (>=1.2.7,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false